+++ Scenes/World/Camera/Shifts/Punch.cs
@@ -8,7 +8,7 @@ public class Punch(Vector2 dir, double strength, double movementSpeed = 3000) : 
     public double Strength { get; private set; } = strength;
     public double InitialStrength { get; private set; } = strength;
 
-    public Vector2 Shift => Direction * Strength;
+    public Vector2 Shift => Direction * (float) Strength;
     public bool IsAlive => Strength > Mathf.Epsilon;
     public void Update(double delta)
     {
